coercing values on chef
Chef has a notion of coercing manifest values similar to Puppet. Chef
will run .parse on every property in the Chef manifest on run. This
ensures that every object is the result of a
Google::product::Property.parse. This won't have any material difference
for now, but will ensure that all nested objects are of a
Google::product::Data type (and resourceref eventually)

Change-Id: Ifb18a9f30082eaf8c0dc93b3a557985c75871611