refactoring chef resourceref to mimic puppet
This refactors Chef ResourceRefs into a separate class to better match
Puppet.

Chef immediately coerces the actual catalog value (including in nested
objects!) into a Google::Data property. The Google::Data property will
either return the immediate value (because it came from an API Request)
or access the global Chef run context and retrieve the value from the
proper block.

This does require that Chef objects are arranged in their proper order
(the referenced block is placed before the block with the
ResourceRef)

Change-Id: I476b064dc28563d91fe85d683695bf3824a5492a